pg2es_hybrid/requirements.txt
llama-index
llama-index-embeddings-ollama
sqlalchemy
elasticsearch
pandas
pyyaml
psycopg2-binary
numpy 